requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
gunicorn
flask-login
werkzeug
pypdf
requests
python-dotenv
Flask-WTF
email_validator
Flask-Migrate
reportlab
google-cloud-aiplatform
thefuzz
google-cloud-storage